Make Writeup changes
After feedback from @FSUMitch & @noproto , made a few changes to make writeup more concise and logically. Also changed names of solution.md to README.md so it would show automatically on github when on that page. Made a seperate file folder that keeps clutter at a minimum when looking at file and updated the relative links in the solution